+++ src/briefy/leica/models/customer/address.py
@@ -15,6 +15,13 @@ class CustomerBillingAddress(Address, mixins.LeicaMixin, Base):
 
     _workflow = workflows.BillingAddressWorkflow
 
+    __summary_attributes__ = [
+        'id', 'created_at', 'updated_at', 'state', 'timezone',
+        'locality', 'country', 'coordinates', 'formatted_address'
+    ]
+
+    __listing_attributes__ = __summary_attributes__
+
     customer_id = sa.Column(
         UUIDType(binary=False),
         sa.ForeignKey('customers.id'), unique=False,