+++ src/briefy/leica/vocabularies/__init__.py
@@ -14,7 +14,7 @@ options = [
 ]
 
 
-ClientJobStatusChoices = LabeledEnum('ClientJobStatusChoices', options)
+OrderStatusChoices = LabeledEnum('OrderStatusChoices', options)
 
 
 options = [
@@ -59,7 +59,7 @@ options = [
 ]
 
 
-JobInputSource = LabeledEnum('JobInputSource', options)
+OrderInputSource = LabeledEnum('OrderInputSource', options)
 
 
 options = [
@@ -69,3 +69,11 @@ options = [
 
 
 ContactTypes = LabeledEnum('ContactTypes', options)
+
+options = [
+    ('new', 'new', 'New Set'),
+    ('returned_photographer', 'returned_photographer', 'Returned by Photographer'),
+    ('refused_customer', 'refused_customer', 'Refused by Customer'),
+]
+
+TypesOfSetChoices = LabeledEnum('TypesOfSetChoices', options)